+++ problems/63.unique-paths-ii.cpp
@@ -1,0 +1,20 @@
+// @leetcode id=63 questionId=63 slug=unique-paths-ii lang=cpp site=leetcode.com title="Unique Paths II"
+class Solution {
+public:
+    int uniquePathsWithObstacles(vector<vector<int>>& obstacleGrid) {
+        int m = obstacleGrid.size(), n = obstacleGrid[0].size();
+        vector<long long> dp(n, 0);
+        dp[0] = (obstacleGrid[0][0] == 0) ? 1 : 0;
+
+        for (int r = 0; r < m; ++r) {
+            for (int c = 0; c < n; ++c) {
+                if (obstacleGrid[r][c] == 1) {
+                    dp[c] = 0;
+                } else if (c > 0) {
+                    dp[c] += dp[c - 1];
+                }
+            }
+        }
+        return (int)dp[n - 1];
+    }
+};